low level support
paradox broken, needs rework on transpose to ice
transpose to ice is also broken.
thunder needs to check, not using it in the begining.
ice stage sometimes stuck because of paradox can't be used.
fire stage is stuck because of paradox can't be used.
AOE not tested.
main fix =>
paradox, transpose